requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.7=py38ha9d4c09_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=py38h06a4308_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.68=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
freetype=2.12.1=h4a9f257_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.4=py38h06a4308_0
joblib=1.4.2=py38h06a4308_0
jpeg=9e=h5eee18b_3
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcurand=10.3.7.68=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.24.3=py38h14f4228_0
numpy-base=1.24.3=py38h31eccc5_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=1.1.1w=h7f8727e_0
packaging=24.1=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
pooch=1.7.0=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.9.0post0=py38h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.4.1=py3.8_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
pyyaml=6.0.1=py38h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=py38h06a4308_0
scikit-learn=1.3.0=py38h1128e8f_1
scipy=1.10.1=py38h14f4228_0
setuptools=74.1.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=py38h06a4308_0
threadpoolctl=3.5.0=py38h2f386ee_0
tk=8.6.14=h39e8969_0
torchaudio=2.4.1=py38_cu121
torchtriton=3.0.0=py38
torchvision=0.19.1=py38_cu121
tqdm=4.66.5=pyhd8ed1ab_0
typing_extensions=4.11.0=py38h06a4308_0
urllib3=2.2.2=py38h06a4308_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
